[Dy2St] Remove `variable_trans_func`
