+++ python/paddle/jit/dy2static/convert_operators.py
@@ -23,8 +23,10 @@ from paddle.autograd.py_layer import PyLayerMeta
 from paddle.base.data_feeder import convert_dtype
 from paddle.base.dygraph.base import _convert_into_variable, in_to_static_mode
 from paddle.base.framework import Variable, core, default_main_program
+from paddle.framework import use_pir_api
 from paddle.pir import Value
 from paddle.static.amp.fp16_utils import AmpOptions
+from paddle.utils import is_sequence, map_structure
 
 from .py_layer import StaticPyLayer
 from .utils import (
@@ -32,10 +34,30 @@ from .utils import (
     Dygraph2StaticException,
     GetterSetterHelper,
     UndefinedVar,
+    create_undefined_variable,
 )
-from .variable_trans_func import to_static_variable
 
 __all__ = []
+
+
+def to_static_variable(x):
+    '''
+    Translate a Python Tensor to PaddlePaddle static graph Tensor
+    '''
+    if isinstance(x, bool):
+        return paddle.full(shape=[], dtype='bool', fill_value=x)
+    if isinstance(x, float):
+        return paddle.full(shape=[], dtype='float64', fill_value=x)
+    if isinstance(x, int):
+        return paddle.full(shape=[], dtype='int64', fill_value=x)
+    if not use_pir_api() and (isinstance(x, UndefinedVar) or x is None):
+        """
+        for early return case, we need a variable to represent None, current we use data_layer_not_check.
+        """
+        return create_undefined_variable()
+    if is_sequence(x):
+        return map_structure(to_static_variable, x)
+    return x
 
 
 def convert_attr(x, attr):
