+++ python/paddle/jit/dy2static/transformers/return_transformer.py
@@ -19,11 +19,10 @@ from ..utils import (
     ORIGI_INFO,
     Dygraph2StaticException,
     ast_to_source_code,
-    index_in_list,
 )
 from .base import BaseTransformer
 from .break_continue_transformer import ForToWhileTransformer
-from .utils import create_bool_node
+from .utils import create_bool_node, index_in_list
 
 __all__ = []
