--- python/paddle/jit/dy2static/variable_trans_func.py
@@ -1,41 +1,0 @@
-#   Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import paddle
-from paddle.framework import use_pir_api
-from paddle.utils import is_sequence, map_structure
-
-from .utils import UndefinedVar, create_undefined_variable
-
-__all__ = []
-
-
-def to_static_variable(x):
-    '''
-    Translate a Python Tensor to PaddlePaddle static graph Tensor
-    '''
-    if isinstance(x, bool):
-        return paddle.full(shape=[], dtype='bool', fill_value=x)
-    if isinstance(x, float):
-        return paddle.full(shape=[], dtype='float64', fill_value=x)
-    if isinstance(x, int):
-        return paddle.full(shape=[], dtype='int64', fill_value=x)
-    if not use_pir_api() and (isinstance(x, UndefinedVar) or x is None):
-        """
-        for early return case, we need a variable to represent None, current we use data_layer_not_check.
-        """
-        return create_undefined_variable()
-    if is_sequence(x):
-        return map_structure(to_static_variable, x)
-    return x
